Fix rotated uv calculations
- Accumulate the rotated angle to determine the uv rotation to use.
- Use the final rotation direction as the block face for uv setting.

+++ polymer/asset/parsed_block_model.h
@@ -25,7 +25,7 @@ struct ParsedRenderableFace {
   Vector2f uv_from;
   Vector2f uv_to;
 
-  float rotation;
+  int rotation;
 
   char texture_name[64];
   size_t texture_name_size;
